+++ src/ir/int.rs
@@ -4,7 +4,7 @@ use crate::{TAPE_LENGTH, ir::ir::{IR, IROp}};
 
 pub fn exec_from_ir(ir: &[IR], offset: u8) {
     let mut pc = 0;
-    let mut offset = offset as isize;
+    let mut offset = offset as i16;
     let mut memory = [0u8; TAPE_LENGTH];
     let mut stdin = stdin().lock();
     let mut stdout = stdout().lock();
